% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch_long}
\alias{ch_long}
\title{Capture Histories - Long Format}
\format{
A data frame with 52 rows and 6 variables:
\describe{
  \item{tag_id}{tag ID including frequency and code}
  \item{loc}{observation location, site or receiver}
  \item{first_obs}{the date and time of first observation at loc}
  \item{last_obs}{the date and time of last observation at loc}
  \item{n}{the number of records that occurred between \code{first_obs} and \code{last_obs}}
  \item{week}{the week number the first observation is from}
}
}
\source{
Biomark NAS - data/telemetry/lemhi/fixed_site_downloads/2018_2019/
}
\usage{
ch_long
}
\description{
Capture histories, in long format, for 5 receivers and tags
in \code{tag_df} from Lemhi River juvenile Chinook salmon winter
telemetry project
}
\keyword{datasets}
