% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch_wide}
\alias{ch_wide}
\title{Capture Histories - Wide Format}
\format{
A data frame with 44 rows and 7 variables:
\describe{
  \item{tag_id}{tag ID including frequency and code}
  \item{cap_hist}{the concatenated capture history}
  \item{LH}{the site's code; Lemhi Hole}
  \item{CA}{the site's code; Carmen Creek}
  \item{TR}{the site's code; Tower Rock}
  \item{RR}{the site's code; Red Rock}
  \item{NF}{the site's code; North Fork}
}
}
\source{
Biomark NAS - data/telemetry/lemhi/fixed_site_downloads/2018_2019/
}
\usage{
ch_wide
}
\description{
Capture histories, in wide format, for 5 receivers and tags
in \code{tag_df} from Lemhi River juvenile Chinook salmon winter
telemetry project
}
\keyword{datasets}
