% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{clean.raw.data}
\alias{clean.raw.data}
\title{Clean raw RT observations}
\usage{
clean.raw.data(raw_data = NULL, filter_valid = T)
}
\arguments{
\item{raw_data}{data.frame containing raw data, output from \code{read.txt.data()}}

\item{filter_valid}{should only valid records (\code{valid == 1}) be returned? Default is \code{TRUE}}
}
\value{
a data frame containing a summary of the raw data
}
\description{
Clean up raw data, including formating dates, deleting invalid rows, etc.
}
\author{
Kevin See
}
