% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayes_cjs.R
\name{fit_bayes_cjs}
\alias{fit_bayes_cjs}
\title{Fit Bayesian CJS Model}
\usage{
fit_bayes_cjs(
  file_path = NULL,
  cap_hist_wide = NULL,
  tag_meta = NULL,
  drop_col_nm = "duty_cycle",
  drop_values = c("batch_2", "batch_3"),
  n_chains = 4,
  n_adapt = 1000,
  n_burnin = 2500,
  n_iter = 2500,
  n_thin = 5,
  params_to_save = c("phi", "p", "survship"),
  rng_seed = 4,
  ...
)
}
\arguments{
\item{file_path}{name (with file path) to save the model as. Probably a .txt file}

\item{cap_hist_wide}{wide capture history, one row per tag, one column per detection site. Includes columns named \code{tag_id} and \code{cap_hist}. Part of the output from \code{prep_capture_history}.}

\item{tag_meta}{metadata for each tag, including a column named \code{tag_id}.}

\item{drop_col_nm}{name of column in \code{tag_meta} to be used to filter out any particular tags.}

\item{drop_values}{character vector of values in \code{drop_col_nm} column of \code{tag_meta} that should be excluded from the CJS model.}

\item{n_chains}{the number of parallel chains for the model}

\item{n_adapt}{the number of iterations for adaptation. If n_adapt = 0 then no adaptation takes place.}

\item{n_burnin}{the number of iterations of the Markov chain to run during the burn-in phase}

\item{n_iter}{the number of iterations to monitor}

\item{n_thin}{the thinning interval for monitors}

\item{params_to_save}{a character vector giving the names of variables to be monitored}

\item{rng_seed}{random number generator seed, to make results reproducible}

\item{...}{other parameters to be fed to \code{rjags::jags.model, update.jags} or \code{rjags::coda.samples}}
}
\value{
tibble
}
\description{
This wrapper function writes a text file describing a Bayesian Cormack Jolly-Seber model in the JAGS language, prepares the data, runs the MCMC and summarises the results.
}
\author{
Kevin See and Mike Ackerman
}
