% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_bad_dates.R
\name{fix_bad_dates}
\alias{fix_bad_dates}
\title{Fix bad dates}
\usage{
fix_bad_dates(detect_df = NULL)
}
\arguments{
\item{detect_df}{data.frame of raw text records, as returned by \code{read_txt_data()}.}
}
\value{
a data frame matching the original, with corrected dates and a new column, \code{orig_date}, showing the original date (character) from the file.
}
\description{
This function corrects those dates at the beginning of a raw text file that are incorrect because the timer has not been reset
}
\author{
Kevin See and Mike Ackerman
}
