% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_nms.R
\name{get.file.nms}
\alias{get.file.nms}
\title{Get File Names}
\usage{
get.file.nms(path = ".")
}
\arguments{
\item{path}{the directory containing the folders with observation data for each telemetry receiver for a study season}
}
\value{
a data frame containing all the files within each telemetry receiver's folder
}
\description{
Generates a data frame containing all the files within each telemetry receiver's folder
}
\author{
Kevin See and Mike Ackerman
}
