% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_alive_vec.R
\name{known_alive_vec}
\alias{known_alive_vec}
\title{Known Alive Vector}
\usage{
known_alive_vec(x_i)
}
\arguments{
\item{xi}{vector of 0's and 1's from a detection history}
}
\value{
a vector containing 1's and NA's
}
\description{
Fill in vector with 1's at points where individual was known to be alive, and NA's elsewhere
}
\author{
Kevin See and Mike Ackerman
}
