% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code_ending.R
\name{parse_code_ending}
\alias{parse_code_ending}
\title{Parse Data Using a Code "Ending"}
\usage{
parse_code_ending(data_to_parse = NULL, code_ending = NULL)
}
\arguments{
\item{data_to_parse}{a data frame to be "parsed" or filtered containing a column \code{tag_id}}

\item{code_ending}{the ending of the tag code to filter for. Some examples include: "575$" for timer tags,
"995$" for noise tags, or "525$" for sentinel tags}
}
\value{
a data frame containing only records for the desired \code{code_ending}
}
\description{
Filter a data frame to extract observations for tag codes ending in \code{code_ending};
useful for extracting noise, timer, etc information.
}
\author{
Kevin See and Mike Ackerman
}
