% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_noise.R
\name{parse_noise}
\alias{parse_noise}
\title{Parse noise data}
\usage{
parse_noise(summ_data = NULL, noise_code = "995$")
}
\arguments{
\item{summ_data}{data.frame containing summarised data, output from \code{summarise_txt_data()}}

\item{noise_code}{code for the tag ID that distinguishes the noise data}
}
\value{
a data.frame containing a summary of the noise data
}
\description{
Filter the summarised data to extract observations related to noise
}
\author{
Kevin See and Mike Ackerman
}
