% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tag_list.R
\name{parse_tag_list}
\alias{parse_tag_list}
\title{Parse Data Using a Tag List}
\usage{
parse_tag_list(data_to_parse = NULL, tags = NULL)
}
\arguments{
\item{data_to_parse}{a data frame to be "parsed" or filtered containing a column \code{tag_id}}

\item{tags}{a vector of tag IDs that the user would like to extract data for}
}
\value{
a data frame containing only records with a \code{tag_id} \%in\% \code{tags}
}
\description{
Filter a data frame to extract observations for a list tags defined by \code{tags}
}
\author{
Kevin See and Mike Ackerman
}
