% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_csv_df.R
\docType{data}
\name{pilot_csv_df}
\alias{pilot_csv_df}
\title{Pilot Study Raw Receiver Data - .csv format}
\format{
csv exports from Tracker software
}
\usage{
pilot_csv_df
}
\description{
The "raw" receiver download data, .csv format, from the Lemhi River radio telemetry pilot study
}
\keyword{csv}
\keyword{data}
\keyword{pilot}
\keyword{radio}
\keyword{telemetry}
