% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_raw.R
\docType{data}
\name{pilot_raw}
\alias{pilot_raw}
\title{Pilot Study Raw Receiver Data - .txt format}
\format{
txt exports from Tracker software
}
\usage{
pilot_raw
}
\description{
The "raw" receiver download data, .txt format, from the Lemhi River radio telemetry pilot study
}
\keyword{data}
\keyword{pilot}
\keyword{radio}
\keyword{telemetry}
\keyword{txt}
