% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_summ.R
\docType{data}
\name{pilot_summ}
\alias{pilot_summ}
\title{Pilot Study summarised Receiver Data - .txt format}
\format{
data.frame from summarise_txt_data
}
\usage{
pilot_summ
}
\description{
The cleaned and summarised receiver data, based on the .txt format, from the Lemhi River radio telemetry pilot study
}
\keyword{data}
\keyword{pilot}
\keyword{radio}
\keyword{telemetry}
