% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_txt_df.R
\docType{data}
\name{pilot_txt_df}
\alias{pilot_txt_df}
\title{Pilot Study Raw Receiver Data - .txt format}
\format{txt exports from Tracker software}
\usage{
pilot_txt_df
}
\description{
The "raw" receiver download data, .txt format, from the Lemhi River radio telemetry pilot study
}
\keyword{data}
\keyword{pilot}
\keyword{radio}
\keyword{telemetry}
\keyword{txt}
