% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs_movement.R
\name{plot_obs_movement}
\alias{plot_obs_movement}
\title{Plot Movement of Observed Tags}
\usage{
plot_obs_movement(ch_long = NULL, receiver_codes = NULL)
}
\arguments{
\item{ch_long}{A long form capture history, produced by the function \code{prep_capture_history()}}

\item{receiver_codes}{character vector of receiver codes to download files for.
The default is \code{NULL}, which will keep all receiver codes found in the path folder}
}
\value{
plot of when each tag was first detected at each receiver
}
\description{
Plots movement through time based on first observations of tags at each receiver. Tags that were never detected after release are not included.
}
\author{
Mike Ackerman and Kevin See
}
