% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volt_temp_data.R
\name{plot_volt_temp_data}
\alias{plot_volt_temp_data}
\title{Plot Receiver Volt and Temperature Info from Tracker Software}
\usage{
plot_volt_temp_data(volt_temp_df, column = "volt_avg", receiver_codes = NULL)
}
\arguments{
\item{volt_temp_df}{A data frame from \code{read_volt_temp_data()}}

\item{column}{A column from volt_temp_df to plot for each receiver. Options from \code{read_volt_temp_data()} include
"volt_avg", "volt_min", "volt_max", "temp_avg", "temp_min", "temp_max".}

\item{receiver_codes}{character vector of receiver codes to download files for.
The default is \code{NULL}, which will keep all receiver codes found in the path folder.}
}
\value{
time-series plots of volt or temp faceted by receiver
}
\description{
Plots volt and temperature information from the Tracker software
}
\author{
Mike Ackerman and Kevin See
}
