% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_capture_history.R
\name{prep_capture_history}
\alias{prep_capture_history}
\title{Prepare Capture Histories}
\usage{
prep_capture_history(
  compress_df = NULL,
  tag_data_path = "data/prepped/tag_release/lemhi_winter_telemetry_tag_info.xlsx",
  n_obs_valid = 3,
  receiver_nms,
  delete_upstream = T,
  output_format = c("wide", "long")
)
}
\arguments{
\item{compress_df}{compressed observational data. Either the output of \code{read_csv_data()} or \code{read_txt_data()} and \code{compress_raw_data()}.}

\item{tag_data_path}{file path to the file containing all the metadata for all tags.}

\item{n_obs_valid}{minumum number of observations made at a location to be considered valid.}

\item{receiver_nms}{character vector of receiver names, in order, to be used in the capture histories.}

\item{delete_upstream}{should detections that indicate upstream movement be deleted?}

\item{output_format}{Either wide or long format}
}
\value{
a dataframe to be used as capture histories
}
\description{
Generates a data frame containing all the necessary information for capture histories of tagged fish
}
\author{
Kevin See and Mike Ackerman
}
