% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_jags_cjs.R
\name{prep_jags_cjs}
\alias{prep_jags_cjs}
\title{Prepare Bayesian CJS Model}
\usage{
prep_jags_cjs(
  cap_hist_wide = NULL,
  tag_meta = NULL,
  drop_col_nm = "duty_cycle",
  drop_values = c("batch_2", "batch_3")
)
}
\arguments{
\item{cap_hist_wide}{wide capture history, one row per tag, one column per detection site. Includes columns named \code{tag_id} and \code{cap_hist}. Part of the output from \code{prep_capture_history}.}

\item{tag_meta}{metadata for each tag, including a column named \code{tag_id}.}

\item{drop_col_nm}{name of column in \code{tag_meta} to be used to filter out any particular tags.}

\item{drop_values}{character vector of values in \code{drop_col_nm} column of \code{tag_meta} that should be excluded from the CJS model.}
}
\value{
list to be fed into JAGS model
}
\description{
Prepares a list of data to be fed into a Bayesian CJS model
}
\author{
Kevin See and Mike Ackerman
}
