% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Raw .txt Format Telemetry Data}
\format{
A data frame with 696930 rows and 9 variables:
\describe{
  \item{file_name}{the full file path}
  \item{file}{the file name}
  \item{date}{date in "\%d\%/\%m\%/\%y\%" format}
  \item{time}{time in "\%H:\%M" format, PST}
  \item{receiver}{receiver code}
  \item{valid}{is observation valid, 1 = T, 0 = F}
  \item{frequency}{channel of record or observation}
  \item{tag_code}{receiver code}
  \item{signal_strength}{receiver code}
}
}
\source{
Biomark NAS - data/telemetry/lemhi/fixed_site_downloads/2018_2019/
}
\usage{
raw
}
\description{
An example dataset containing observations for 5 receivers
and 1 frequency, 2018/2019 season
}
\keyword{datasets}
