% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data.R
\name{read.csv.data}
\alias{read.csv.data}
\title{Read in .csv Format Tracker Data}
\usage{
read.csv.data(path = ".")
}
\arguments{
\item{path}{the directory containing the folders with observation data for each telemetry receiver for a study season}
}
\value{
a data frame containing all of the .csv data
}
\description{
Reads in telemetry observation data stored in the .csv format from the Tracker software
}
\author{
Kevin See
}
