% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_volt_temp_data.R
\name{read.volt.temp.data}
\alias{read.volt.temp.data}
\title{Read in Receiver Volt and Temperature Info from Tracker Software}
\usage{
read.volt.temp.data(path = ".")
}
\arguments{
\item{path}{the directory containing the folders with observation data for each telemetry receiver for a study season}
}
\value{
a data frame containing receiver volt and temperature information
}
\description{
Reads in volt and temperature receiver data from Tracker software
}
\author{
Mike Ackerman and Kevin see
}
