% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_txt_data.R
\name{read_txt_data}
\alias{read_txt_data}
\title{Read in raw .txt Format Tracker Data}
\usage{
read_txt_data(path = ".", receiver_codes = NULL)
}
\arguments{
\item{path}{the directory containing the folders with observation data for each telemetry receiver for a study season}

\item{receiver_codes}{character vector of receiver codes to query for. Default is \code{NULL} which will keep all receiver codes found in the path folder}
}
\value{
a data frame containing all of the .txt data
}
\description{
Reads in raw telemetry observation data stored in the .txt format from the Tracker software
}
\author{
Kevin See and Mike Ackerman
}
