% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_tag_codes.R
\name{round.tag.codes}
\alias{round.tag.codes}
\title{Round Tag Codes}
\usage{
\method{round}{tag.codes}(data = NULL, round_to = 10)
}
\arguments{
\item{data}{data.frame with column named \code{tag_code}}

\item{round_to}{integer that the tag codes should be rounded to, currently supports either 10 or 5 (default is 10)}
}
\value{
a data frame containing all the original columns, plus one called \code{tag_id}
}
\description{
Round the tag IDs from the receiver to better match possible tag codes
}
\author{
Kevin See and Mike Ackerman
}
