% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags_cjs.R
\name{run_jags_cjs}
\alias{run_jags_cjs}
\title{Run Bayesian CJS Model}
\usage{
run_jags_cjs(
  file_path = NULL,
  jags_data = NULL,
  n_chains = 4,
  n_adapt = 1000,
  n_burnin = 2500,
  n_iter = 2500,
  n_thin = 5,
  params_to_save = c("phi", "p", "survship"),
  rng_seed = 4
)
}
\arguments{
\item{file_path}{name (with file path) to save the model as. Probably a .txt file}

\item{jags_data}{list of data to be fed into the JAGS CJS model. Generated from \code{prep_jags_cjs}.}

\item{n_chains}{the number of parallel chains for the model}

\item{n_adapt}{the number of iterations for adaptation. If n_adapt = 0 then no adaptation takes place.}

\item{n_burnin}{the number of iterations of the Markov chain to run during the burn-in phase}

\item{n_iter}{the number of iterations to monitor}

\item{n_thin}{the thinning interval for monitors}

\item{params_to_save}{a character vector giving the names of variables to be monitored}

\item{rng_seed}{random number generator seed, to make results reproducible}

\item{...}{other parameters to be fed to \code{rjags::jags.model, rjags::update} or \code{rjags::coda.samples}}
}
\value{
mcmc.list
}
\description{
Generate MCMC samples from the posteriors of a Bayesian Cormack Jolly-Seber model.
}
\author{
Kevin See and Mike Ackerman
}
