% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_1819_csv_df.R
\docType{data}
\name{ssn_1819_csv_df}
\alias{ssn_1819_csv_df}
\title{Raw Receiver Data from 2018/2019 season - .csv format}
\format{csv exports from Tracker software}
\usage{
ssn_1819_csv_df
}
\description{
The "raw" receiver download data, .csv format, from the Lemhi River radio telemetry study, 2018-2019 season
}
\keyword{csv}
\keyword{data}
\keyword{radio}
\keyword{telemetry}
