% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_1819_txt_df.R
\docType{data}
\name{ssn_1819_txt_df}
\alias{ssn_1819_txt_df}
\title{Raw Receiver Data from 2018/2019 Season - .txt format}
\format{
txt exports from Tracker software
}
\usage{
ssn_1819_txt_df
}
\description{
The "raw" receiver download data, .txt format, from the Lemhi River radio telemetry study, 2018-2019 season
}
\keyword{data}
\keyword{radio}
\keyword{telemetry}
\keyword{txt}
