% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn_1920_raw.R
\docType{data}
\name{ssn_1920_raw}
\alias{ssn_1920_raw}
\title{Raw Receiver Data from 2019/2020 Season - .txt format}
\format{
txt exports from Tracker software
}
\usage{
ssn_1920_raw
}
\description{
The "raw" receiver download data, .txt format, from the Lemhi River radio telemetry study, 2019-2020 season
}
\keyword{data}
\keyword{radio}
\keyword{telemetry}
\keyword{txt}
