% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_txt_data.R
\name{summarise.txt.data}
\alias{summarise.txt.data}
\title{Summarise raw RT observations}
\usage{
\method{summarise}{txt.data}(data = NULL)
}
\arguments{
\item{data}{data.frame containing all valid observations, output from \code{read.txt.data()}, followed by \code{clean.raw.data()}, followed by \code{round.tag.codes()}}
}
\value{
a data frame containing a summary of the raw data
}
\description{
Summarises cleaned detections, showing the first and last detection on each receiver for each tag, and how many times it was detected on that receiver in that window.
}
\author{
Kevin See
}
