% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_jags_cjs.R
\name{summarise_jags_cjs}
\alias{summarise_jags_cjs}
\title{Summarise Bayesian CJS Model}
\usage{
summarise_jags_cjs(post = NULL, p = NULL, ...)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p}{a character vector of with length >= 1 specifying the nodes to summarize.
Passed to \code{\link[postpack]{match_p}}, so can (and sometimes should) be a regular expression.}
}
\value{
tibble
}
\description{
From the mcmc.list object, summarise the parameter estimates and include site names where appropriate.
}
\author{
Kevin See and Mike Ackerman
}
