% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_noise_data.R
\name{summarise_noise_data}
\alias{summarise_noise_data}
\title{Summarize Receiver Noise Information}
\usage{
summarise_noise_data(
  compress_df = NULL,
  receiver_codes = NULL,
  operations_summary = NULL
)
}
\arguments{
\item{compress_df}{compressed observational data. Either the output of \code{read_csv_data()} or \code{read_txt_data()} and \code{compress_raw_data()}.}

\item{receiver_codes}{character vector of receiver codes to download files for.
The default is \code{NULL}, which will keep all receiver codes found in the path folder}

\item{operations_summary}{an optional summary of receiver operation times typically the \code{operations_summ} object
from \code{summarise_timer_data()}}
}
\value{
summaries of noise information
}
\description{
Summarize receiver noise records in observation data
}
