% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_test_data.R
\name{summarise_test_data}
\alias{summarise_test_data}
\title{Summarize Test Tag Information}
\usage{
summarise_test_data(compress_df = NULL, tag_data = NULL)
}
\arguments{
\item{compress_df}{compressed observational data. Either the output of \code{read_csv_data()} or \code{read_txt_data()} and \code{compress_raw_data()}.}

\item{tag_data}{dataframe with metadata for each tag, including columns named \code{tag_id}, \code{tag_purpose} (with some tags having "\code{fish}" in this column), and \code{release_time}.}
}
\value{
summaries of test tag data
}
\description{
Summarize test tag information including battery life
}
\author{
Mike Ackerman and Kevin See
}
