% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_timer_data.R
\name{summarise_timer_data}
\alias{summarise_timer_data}
\title{Summarize Receiver Operation Times}
\usage{
summarise_timer_data(
  compress_df = NULL,
  receiver_codes = NULL,
  season_start = NULL,
  season_end = NULL,
  include_noise = c(TRUE, FALSE)
)
}
\arguments{
\item{compress_df}{compressed observational data. Either the output of \code{read_csv_data()} or \code{read_txt_data()} and \code{compress_raw_data()}.}

\item{receiver_codes}{character vector of receiver codes to download files for.
The default is \code{NULL}, which will keep all receiver codes found in the path folder}

\item{season_start}{character with format "YYYYMMDD" the first day of the season that the receivers were considered turned on}

\item{season_end}{character with format "YYYYMMDD" the final day of the season that the receivers were all turned off}

\item{include_noise}{would user like to include \code{TRUE} or not \code{FALSE} any
noise records ending in "995" when summarising receiver operational times. Default
is \code{TRUE}.}
}
\value{
summaries of operational times for each receiver
}
\description{
Summarize receiver operational times using data from the timer tags at each site
}
\author{
Mike Ackerman and Kevin See
}
