% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_txt_data.R
\name{summarise_txt_data}
\alias{summarise_txt_data}
\title{Summarise raw RT observations}
\usage{
summarise_txt_data(data_df = NULL, max_min = 5)
}
\arguments{
\item{data_df}{data.frame containing all valid observations, output from \code{read_txt_data()}, followed by \code{clean_raw_data()}, followed by \code{round_tag_codes()}}

\item{max_min}{maximum number of minutes between detections of a tag before it's considered a different "group" of detections. Default is 5.}
}
\value{
a data.frame containing a summary of the raw data
}
\description{
Summarises cleaned detections, showing the first and last detection on each receiver for each tag, and how many times it was detected on that receiver in that window.
}
\author{
Kevin See and Mike Ackerman
}
