% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tag_df}
\alias{tag_df}
\title{Tag Releases - For Capture Histories}
\format{
A data frame with 38 rows and 13 variables:
\describe{
  \item{season}{the year/season the tag was deployed}
  \item{radio_tag_id}{the radio tag ID including frequency, code and letter}
  \item{pit_tag_id}{the PIT tag number also injected into the fish}
  \item{tag_id}{tag ID including frequency and code}
  \item{srr}{species, run, rear, uses PTAGIS nomenclature}
  \item{length}{length, millimeters}
  \item{weight}{weight, grams}
  \item{tag_purpose}{Was tag deployed in a fish, used as test tag, or other?}
  \item{release_site}{location that a fish was released at}
  \item{duty_cycle}{the batch or duty cycle of the fish}
  \item{activation_time}{the activation time of the tag, Pacific Standard}
  \item{release_time}{the release time for the fish, Pacific Standard}
  \item{notes}{relevant notes}
}
}
\source{
Biomark NAS - data/telemetry/lemhi/fixed_site_downloads/2018_2019/
}
\usage{
tag_df
}
\description{
A subset of \code{tag_releases} for those tags included
in \code{ch_long} and \code{ch_wide}
}
\keyword{datasets}
