% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tag_releases}
\alias{tag_releases}
\title{Example Tag Release Data}
\format{
A data frame with 38 rows and 13 variables:
\describe{
  \item{season}{the year/season the tag was deployed}
  \item{radio_tag_id}{tag ID including frequency, code, and letter denoting season}
  \item{pit_tag_id}{the PIT tag number also injected into the fish}
  \item{tag_id}{\code{radio_tag_id}, but with season letter trimmed off}
  \item{srr}{species, run, rear, uses PTAGIS nomenclature}
  \item{length}{length, millimeters}
  \item{weight}{weight, grams}
  \item{tag_purpose}{Was tag deployed in a fish, used as test tag, or other?}
  \item{release_site}{location that a fish was released at}
  \item{duty_cycle}{the batch or duty cycle of the fish}
  \item{activation_time}{the activation time of the tag, Pacific Standard}
  \item{release_time}{the release time for the fish, Pacific Standard}
  \item{notes}{relevant notes}
}
}
\source{
Biomark NAS - data/telemetry/lemhi/tag_release/
}
\usage{
tag_releases
}
\description{
Example tag release data; \code{tag_releases} includes information
from 38 tags released in fall 2018 in a juvenile Chinook salmon
radio telemetry project on the Lemhi and Salmon Rivers, Idaho
}
\keyword{datasets}
