% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volt_temp_df.R
\docType{data}
\name{volt_temp_df}
\alias{volt_temp_df}
\title{Volt & Temperature Data}
\format{

}
\source{
Biomark NAS - data/telemetry/lemhi/fixed_site_downloads/2018_2019/
}
\usage{
volt_temp_df
}
\description{
An example dataset containing voltage and temperature data recorded
by 3 telemetry receivers, 2018/2019 season
}
\keyword{datasets}
