% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bayes_cjs.R
\name{write_bayes_cjs}
\alias{write_bayes_cjs}
\title{Write Bayesian CJS Model}
\usage{
write_bayes_cjs(file_path = NULL)
}
\arguments{
\item{file_path}{name (with file path) to save the model as. Probably a .txt file}
}
\description{
Creates a text file describing a Bayesian Cormack Jolly-Seber model.
}
\author{
Kevin See and Mike Ackerman
}
