% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_raw_data.R
\name{compress_raw_data}
\alias{compress_raw_data}
\title{Compress raw RT observations}
\usage{
compress_raw_data(
  data_df = NULL,
  filter_valid,
  round_to,
  max_min,
  assign_week,
  week_base,
  append_week
)
}
\arguments{
\item{data_df}{data.frame with column named \code{tag_code}}

\item{filter_valid}{should only valid records (\code{valid == 1}) be returned? Default is \code{TRUE}}

\item{round_to}{integer that the tag codes should be rounded to, currently supports either 10 or 5 (default is 10)}

\item{max_min}{maximum number of minutes between detections of a tag before it's considered a different "group" of detections. Default is 2.}

\item{assign_week}{Should this function assign a week number to the output? Default is \code{TRUE}}

\item{week_base}{If assigning week numbers, the date when the numbering should start in MMDD format}

\item{append_week}{If assigning weeks, should the week be assigned based on the \code{first} or \code{last} time a tag was detected on that receiver? Default value is \code{first}.}
}
\value{
a data.frame containing a summary of the raw data
}
\description{
Cleans detections, rounds observed tag codes and compresses the detections, showing the first and last detection on each receiver for each tag, and how many times it was detected on that receiver in that window.
}
\author{
Kevin See and Mike Ackerman
}
