% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_raw_data.R
\name{compress_raw_data}
\alias{compress_raw_data}
\title{Compress raw RT observations}
\usage{
compress_raw_data(
  data_df = NULL,
  min_yr = 2017,
  max_yr = NA,
  filter_valid = T,
  round_to = 10,
  max_min = 2,
  assign_week = T,
  week_base = "0901",
  append_week = c("first", "last")
)
}
\arguments{
\item{data_df}{any data frame with a column named \code{tag_code}}

\item{min_yr}{the minimum acceptable year (e.g. 2017), likely the year at the start of a season. Used to filter
errant dates that occur within the data pre-season.}

\item{max_yr}{maximum acceptable year (e.g. 2018), likely the year at the end of a season. Used to filter errant
dates that occur within the data post-season.}

\item{filter_valid}{should only valid records in the \code{raw_data} be returned?
Default is \code{TRUE} which only returns records where \code{valid == 1}. This uses the 4th column in the
raw Tracker software .txt downloads.}

\item{round_to}{the integer that the user would like to round \code{tag_code} to. Currently supports
either 5 or 10. If \code{round_to = 5}, will round to the nearest 5. If \code{round_to = 10}, will round
to the nearest 0.}

\item{max_min}{maximum number of minutes between detections of a tag before it's considered a different "group" of detections. Default is 2.}

\item{assign_week}{Should this function assign a week number to the output? Default is \code{TRUE}}

\item{week_base}{If assigning week numbers, the date when the numbering should start in MMDD format}

\item{append_week}{If assigning weeks, should the week be assigned based on the \code{first} or \code{last} time a tag was detected on that receiver? Default value is \code{first}.}
}
\value{
a data.frame containing a summary of the raw data
}
\description{
Cleans detections, rounds observed tag codes and compresses the detections, showing the first and last detection on each receiver for each tag, and how many times it was detected on that receiver in that window.
}
\author{
Kevin See and Mike Ackerman
}
