% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code_ending.R
\name{parse_code_ending}
\alias{parse_code_ending}
\title{Parse tags by tag code ending}
\usage{
parse_code_ending(summ_data = NULL, code_ending = NULL)
}
\arguments{
\item{summ_data}{data.frame containing summarised data, output from \code{summarise_txt_data()}}

\item{code_ending}{the ending of the tag code to filter for. Some examples include: "575$" for timer tags, "995$" for noise tags or "525$" for sentinel tags}
}
\value{
a data.frame containing a only records with the desired code ending
}
\description{
Filter the summarised data to extract observations related to timer tags
}
\author{
Kevin See and Mike Ackerman
}
