% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tag_list.R
\name{parse_tag_list}
\alias{parse_tag_list}
\title{Parse tag data}
\usage{
parse_tag_list(summ_data = NULL, tags = NULL)
}
\arguments{
\item{summ_data}{data.frame containing summarised data, output from \code{summarise_txt_data()}}

\item{tags}{vector of tag IDs that the user would like to extract data for}
}
\value{
a data.frame containing a summary of the tag data for the listed tags
}
\description{
Filter the summarised data to extract observations related to tags, defined in a tag list
}
\author{
Kevin See and Mike Ackerman
}
