% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tag_list.R, R/parse_timer.R
\name{parse_timer}
\alias{parse_timer}
\title{Parse tag data}
\usage{
parse_timer(summ_data = NULL, timer_code = "575$")

parse_timer(summ_data = NULL, timer_code = "575$")
}
\arguments{
\item{summ_data}{data.frame containing summarised data, output from \code{summarise_txt_data()}}

\item{tags}{vector of tag IDs that the user would like to extract data for}

\item{noise_code}{code for the tag ID that distinguishes the timer tag data}
}
\value{
a data.frame containing a summary of the tag data for the listed tags

a data.frame containing a summary of the timer tag data
}
\description{
Filter the summarised data to extract observations related to tags, defined in a tag list

Filter the summarised data to extract observations related to timer tags
}
\author{
Kevin See and Mike Ackerman

Kevin See and Mike Ackerman
}
