% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_capture_history.R
\name{prep_capture_history}
\alias{prep_capture_history}
\title{Prepare Capture Histories}
\usage{
prep_capture_history(
  compress_df = NULL,
  tag_data = NULL,
  n_obs_valid = 3,
  rec_site = NULL,
  delete_upstream = T,
  location = c("site", "receiver"),
  output_format = c("all", "wide", "long")
)
}
\arguments{
\item{compress_df}{compressed observational data. Either the output of \code{read_csv_data()} or \code{read_txt_data()} and \code{compress_raw_data()}.}

\item{tag_data}{dataframe with metadata for each tag, including columns named \code{tag_id}, \code{tag_purpose} (with some tags having "\code{fish}" in this column), \code{release_site} (which is one of the sites in the \code{rec_site} input) and \code{release_time}.}

\item{n_obs_valid}{minumum number of observations made at a location to be considered valid.}

\item{rec_site}{data.frame with 2 columns, \code{site} and \code{receiver}, with sites and and receivers as factors with levels in order.}

\item{delete_upstream}{should detections that indicate upstream movement be deleted?}

\item{location}{should detections be grouped by site or receiver?}

\item{output_format}{return either wide or long format, or both plus tag metadata}
}
\value{
a list containing a capture history in wide format, one in long format, and metadata for each tag
}
\description{
Generates a data frame containing all the necessary information for capture histories of tagged fish
}
\author{
Kevin See and Mike Ackerman
}
