% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_on_off_data.R
\name{read.on.off.data}
\alias{read.on.off.data}
\title{Read in Receiver On/Off Times from Tracker Software}
\usage{
read.on.off.data(path = ".")
}
\arguments{
\item{path}{the directory containing the folders with observation data for each telemetry receiver for a study season}
}
\value{
a data frame containing receiver on/off information
}
\description{
Reads in information on receiver on/off times from the Tracker software
}
\author{
Mike Ackerman and Kevin See
}
