% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data.R
\name{read_csv_data}
\alias{read_csv_data}
\title{Read in .csv Format Tracker Data}
\usage{
read_csv_data(path = ".", receiver_codes = NULL)
}
\value{
a data frame containing all of the .csv data
}
\description{
Reads in telemetry observation data stored in the .csv format from the Tracker software
}
\author{
Kevin See
}
