% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_on_off_data.R
\name{read_on_off_data}
\alias{read_on_off_data}
\title{Read in Receiver On/Off Times from Tracker Software}
\usage{
read_on_off_data(path = ".", receiver_codes = NULL)
}
\arguments{
\item{path}{a path to the directory containing all of the files downloaded from radio telemetry receivers
for a study season and using the Tracker software}

\item{receiver_codes}{character vector of receiver codes to download files for.
The default is \code{NULL}, which will keep all receiver codes found in the path folder}
}
\value{
a data frame containing receiver on/off information
}
\description{
Reads in information on receiver on/off times from the Tracker software
}
\author{
Mike Ackerman and Kevin See
}
