% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_volt_temp_data.R
\name{read_volt_temp_data}
\alias{read_volt_temp_data}
\title{Read in Receiver Volt and Temperature Info from Tracker Software}
\usage{
read_volt_temp_data(path = ".", receiver_codes = NULL)
}
\arguments{
\item{path}{a path to the directory containing all of the files downloaded from radio telemetry receivers
for a study season and using the Tracker software.
using the Tracker Software}

\item{receiver_codes}{character vector of receiver codes to download files for.
The default is \code{NULL}, which will keep all receiver codes found in the path folder.}
}
\value{
a data frame containing receiver volt and temperature information
}
\description{
Reads in volt and temperature data from the Tracker software
}
\author{
Mike Ackerman and Kevin See
}
