% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_metadata}
\alias{site_metadata}
\title{Example Site Metadata}
\format{
A data frame with 5 rows and 12 variables:
\describe{
  \item{site_name}{name of release or observation site}
  \item{site_code}{site code, PTAGIS code in the case of \code{site_type == rst} | \code{site_type == pit_array}}
  \item{site_type}{rotary screw trap (rst), fixed radio telemetry, or PIT tag array}
  \item{receivers}{a list of receivers that have been used at the site}
  \item{rt_site_type}{for \code{site_type == rt_fixed}, pod or trailer}
  \item{point_y}{latitude, WGS84, decimal degrees}
  \item{point_x}{longitude, WGS84, decimal degrees}
  \item{use17_18}{Was the site used during the 2017/2018 pilot season?}
  \item{use18_19}{Was the site used during the 2018/2019 season?}
  \item{use19_20}{Was the site used during the 2019/2020 season?}
  \item{ptagis_rkm}{river rkm from PTAGIS}
  \item{rt_rkm}{river rkm for Lemhi River telemetry project, \code{site_code == TB} was designated as "000"}
}
}
\source{
Biomark NAS - data/telemetry/lemhi/site_metadata/
}
\usage{
site_metadata
}
\description{
Example site metadata; \code{site_metadata} includes information
from 5 sites used in a juvenile Chinook salmon radio telemetry
project on the Lemhi and Salmon Rivers, Idaho
}
\keyword{datasets}
