% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_noise_data.R
\name{summarise_noise_data}
\alias{summarise_noise_data}
\title{Summarize Receiver Noise Information}
\usage{
summarise_noise_data(
  summ_data = NULL,
  receiver_codes = NULL,
  operations_summary = NULL
)
}
\arguments{
\item{summ_data}{data.frame containing summarised data, output from \code{summarise_txt_data()}}

\item{receiver_codes}{character vector of receiver codes to query for. Default is \code{NULL} which will keep all receiver codes found in the path folder}

\item{operations_summary}{an optional summary of receiver operation times typically the \code{operations_summ} object
from \code{summarise_timer_data()}}
}
\value{
summaries of noise information
}
\description{
Summarize receiver noise records in observation data
}
