% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_timer_data.R
\name{summarise_timer_data}
\alias{summarise_timer_data}
\title{Summarize Receiver Operation Times}
\usage{
summarise_timer_data(
  summ_data = NULL,
  receiver_codes = NULL,
  season_start = NULL,
  season_end = NULL
)
}
\arguments{
\item{summ_data}{data.frame containing summarised data, output from \code{summarise_txt_data()}}

\item{receiver_codes}{character vector of receiver codes to query for. Default is \code{NULL} which will keep all receiver codes found in the path folder}

\item{season_start}{format "\%Y-\%m-\%d" the first day of the season that the receivers were considered turned on}

\item{season_end}{format "\%Y-\%m-\%d" the final day of the season that the receivers were all turned off}
}
\value{
summaries of operational times for each receiver
}
\description{
Summarize receiver operational times using data from the timer tags at each site
}
\author{
Mike Ackerman and Kevin See
}
