% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_timer_data.R
\name{summarise_timer_data}
\alias{summarise_timer_data}
\title{Summarize Receiver Operation Times}
\usage{
summarise_timer_data(timer_data = NULL, receiver_codes = NULL)
}
\arguments{
\item{timer_data}{the data frame containing the timer tag data e.g., from parse_timer()}

\item{receiver_codes}{character vector of receiver codes to query for. Default is \code{NULL} which will keep all receiver codes found in the path folder}
}
\value{
a data frame of T/F summarizing operational times for each receiver
}
\description{
Summarize receiver operational times using data from the timer tags at each site
}
\author{
Mike Ackerman and Kevin See
}
